Fix problem when loading user data
Every time the app was reloaded, the logged in user data
disappeared, so I had to look for the root of the problem.
In the end it turned out that I had added an "async" in a
constant that didn't need it, so I removed that part of
the code.